+++ bzt/engine/__init__.py
@@ -1,3 +1,5 @@
 from .names import EXEC, TAURUS_ARTIFACTS_DIR, SETTINGS
-from .engine import Engine, Configuration, FileLister, Scenario, HavingInstallableTools, Singletone, SelfDiagnosable
+from .engine import Engine
+from .dicts import Scenario, Configuration
+from .templates import FileLister, HavingInstallableTools, Singletone, SelfDiagnosable
 from .modules import Provisioning, Reporter, Service, Aggregator, EngineModule, ScenarioExecutor
