+++ bzt/engine/templates.py
@@ -1,0 +1,53 @@
+"""
+BZT templates
+
+Copyright 2019 BlazeMeter Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+from abc import abstractmethod
+
+
+class FileLister(object):
+    """
+    A mixin to get required files info from executor
+    """
+
+    @abstractmethod
+    def resource_files(self):
+        """
+        Get list of resource files
+
+        :rtype: list
+        """
+        pass
+
+
+class HavingInstallableTools(object):
+    @abstractmethod
+    def install_required_tools(self):
+        pass
+
+
+class Singletone(object):
+    pass
+
+
+class SelfDiagnosable(object):
+    @abstractmethod
+    def get_error_diagnostics(self):
+        """
+
+        :rtype: list[str]
+        """
+        pass
